+++ src/sim/status_warehouse/entry/drawer_entry.py
@@ -20,8 +20,15 @@ class DrawerEntry(Entry):
     def __deepcopy__(self, memo):
         copy_obj = DrawerEntry(self.get_offset_x(), self.get_pos_y())
         copy_obj.drawer = copy.deepcopy(self.get_drawer(), memo)
-        copy_obj.drawer.set_first_drawerEntry(copy_obj)
+        if copy_obj == self.get_drawer().get_first_drawerEntry():
+            copy_obj.drawer.set_first_drawerEntry(copy_obj)
         return copy_obj
+
+    def __eq__(self, other):
+        return (
+            isinstance(other, DrawerEntry) and
+            Entry.__eq__(self, other)
+        )
 
     def get_drawer(self) -> Drawer:
         """